% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth}
\alias{auth}
\title{Authenticate to the Evalyzer platform.}
\usage{
auth(username, password)
}
\arguments{
\item{username}{User name.}

\item{password}{User password.}
}
\value{
A rvest session.
}
\description{
Authenticate to the Evalyzer platform.
}
\examples{
session <- auth("my-username", "my-password")
}
