% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_answer.R
\name{get_answer}
\alias{get_answer}
\title{Get the answers of all users}
\usage{
get_answer(session, id)
}
\arguments{
\item{session}{A rvest session. The rvest session is created by the
\code{\link{auth}} function.}

\item{id}{The id of a project. The id lies in url like
\url{https://app.evalyzer.com/evalyzer/analyzes/participants/<id>}}
}
\value{
A \code{data.frame}  with the pre-test, post-task and post-test answers.
}
\description{
Get the pre-test, post-task and post-test answers of all users.
}
\examples{
session <- auth("my-username", "my-password")
participant_answers <- get_answer(session, "1234")
}
