% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_behavior.R
\name{get_behavior}
\alias{get_behavior}
\title{Get behaviors of all users}
\usage{
get_behavior(session, id)
}
\arguments{
\item{session}{A rvest session. The rvest session is created by the}

\item{id}{The id of a project. The id lies in url like
\code{\link{auth}} function.}
}
\value{
A \code{data.frame} with user timestamps, clicks, scrolls, page inputs,
consulted pages.
}
\description{
Get behaviors of all users for each task.
The behaviors mainly include:
\itemize{
\item clicks,
\item scrolls,
\item page inputs
\item consulted pages.
}
}
\examples{
session <- auth("my-username", "my-password")
user_behavior <- get_behavior(session, "project-id")
}
