% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_configuration.R
\name{get_configuration}
\alias{get_configuration}
\title{Get configuration of each user}
\usage{
get_configuration(session, id)
}
\arguments{
\item{session}{A rvest session. The rvest session is created by the
\code{\link{auth}} function.}

\item{id}{The id of a project. The id lies in url like
\url{https://app.evalyzer.com/evalyzer/analyzes/participants/<id>}}
}
\value{
A \code{data.frame} with the screen resolution, browser, operating system
information.
}
\description{
Get configuration of each user for each task.
Configuration information includes:
\itemize{
\item the screen resolution information,
\item the browser information,
\item the operating system information.
}
}
\examples{
session <- auth("my-username", "my-password")
user_configuration <- get_configuration(session, "project-id")
}
