% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_overview.R
\name{get_overview}
\alias{get_overview}
\title{Get an overview of all users}
\usage{
get_overview(session, id)
}
\arguments{
\item{session}{A rvest session. The rvest session is created by the
\code{\link{auth}} function.}

\item{id}{The id of a project. The id lies in url like
\url{https://app.evalyzer.com/evalyzer/analyzes/participants/<id>}}
}
\value{
A \code{data.frame} with the user ids, the dates and times, the task
performances (task duration and task status), the total duration of
the task(s).
}
\description{
Get some data about each user:
* the user id,
* the test session date and time,
* the task performances (task duration and task status),
* the total duration of the task(s).
}
\examples{
session <- auth("my-username", "my-password")
user_overview <- get_overview(session, "project-id")
}
