% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_video.R
\name{get_video}
\alias{get_video}
\title{Get the videos of each user for each task.}
\usage{
get_video(session, id, path)
}
\arguments{
\item{session}{The session is available by the \code{\link{auth}} function.
Please make sure you have used this function beforehand.}

\item{id}{The id of a project. The id lies in url like
\url{https://app.evalyzer.com/evalyzer/analyzes/participation/<id>}}

\item{path}{The path to store videos.}
}
\value{
The user videos.
}
\description{
Get the videos of all participants for each task.
}
\examples{
session <- auth("my-username", "my-password")
get_video(session, "1234", "pathtostorefiles/")
}
